move ShardingTableMetasLoader to sharding-core-common
