+++ sharding-core/sharding-core-common/src/main/java/org/apache/shardingsphere/core/metadata/ShardingTableMetaDataDecorator.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sharding.execute.metadata;
+package org.apache.shardingsphere.core.metadata;
 
 import org.apache.shardingsphere.core.metadata.column.ShardingGeneratedKeyColumnMetaData;
 import org.apache.shardingsphere.core.rule.DataNode;
